Type fix

+++ src/engine/renderer/renderer.h
@@ -57,8 +57,8 @@ class Renderer {
                               const void* indices) = 0;
   virtual void DestroyGeometry(uint64_t resource_id) = 0;
   virtual void Draw(uint64_t resource_id,
-                    uint64_t num_indices = 0,
-                    uint64_t start_offset = 0) = 0;
+                    size_t num_indices = 0,
+                    size_t start_offset = 0) = 0;
 
   virtual uint64_t CreateTexture() = 0;
   virtual void UpdateTexture(uint64_t resource_id,
@@ -70,7 +70,7 @@ class Renderer {
                              size_t data_size,
                              uint8_t* image_data) = 0;
   virtual void DestroyTexture(uint64_t resource_id) = 0;
-  virtual void ActivateTexture(uint64_t resource_id, uint64_t texture_unit) = 0;
+  virtual void ActivateTexture(uint64_t resource_id, size_t texture_unit) = 0;
 
   virtual uint64_t CreateShader(std::unique_ptr<ShaderSource> source,
                                 const VertexDescription& vertex_description,
